// workspaces/mooncg/src/types/mooncg.ts
type Person =
	| {
			name: string;
			email?: string;
			url?: string;
	  }
	| string;

export namespace MoonCG {
	/**
	 * A collection of types that represent the raw data from the `mooncg` stanza in a bundle's `package.json`.
	 */
	export namespace Manifest {
		/**
		 * An unparsed `assetCategory` from a bundle's `package.json`.
		 */
		export interface UnparsedAssetCategory {
			name: string;
			title: string;
			allowedTypes?: string[];
		}

		/**
		 * An unparsed `panel` from a bundle's `package.json`.
		 */
		export interface UnparsedPanel {
			name: string;
			title: string;
			file: string;
			headerColor?: string;
			fullbleed?: boolean;
			workspace?: string;
			dialog?: boolean;
			dialogButtons?: { name: string; type: "dismiss" | "confirm" }[];
			width?: number;
		}

		/**
		 * An unparsed `graphic` from a bundle's `package.json`.
		 */
		export interface UnparsedGraphic {
			file: string;
			width: number;
			height: number;
			singleInstance?: boolean;
		}

		/**
		 * An unparsed `mount` configuration from a bundle's `package.json`.
		 */
		export interface UnparsedMount {
			directory: string;
			endpoint: string;
		}

		/**
		 * An unparsed `soundCue` from a bundle's `package.json`.
		 */
		export interface UnparsedSoundCue {
			name: string;
			assignable?: boolean;
			defaultVolume?: number;
			defaultFile?: string;
		}

		/**
		 * An unparsed record of `bundleDependencies` from a bundle's `package.json`.
		 */
		export type UnparsedBundleDependencies = Record<string, string>;

		/**
		 * The actual structure of a bundle's `mooncg` stanza from it's `package.json`.
		 */
		export interface UnparsedManifest {
			compatibleRange?: string;
			transformBareModuleSpecifiers?: boolean;
			dashboardPanels?: UnparsedPanel[];
			graphics?: UnparsedGraphic[];
			assetCategories?: UnparsedAssetCategory[];
			mount?: UnparsedMount[];
			soundCues?: UnparsedSoundCue[];
			bundleDependencies?: UnparsedBundleDependencies;
		}
	}
// ...
	export type Manifest = Omit<PackageJSON, "mooncg"> &
		Manifest.UnparsedManifest & { transformBareModuleSpecifiers: boolean };

	/**
	 * A collection of types that comprise a `bundle`.
	 */
	export namespace Bundle {
		export type GitData =
			| undefined
			| {
					branch: string;
					hash: string;
					shortHash: string;
			  }
			| {
					branch: string;
					hash: string;
					shortHash: string;
					date: string;
					message: string;
			  };

		export type Graphic = {
			url: string;
		} & Required<Manifest.UnparsedGraphic>;

		export type Panel = Manifest.UnparsedPanel & {
			path: string;
			headerColor: string;
			bundleName: string;
			html: string;
		} & (
				| {
						dialog: false;
						workspace: string;
				  }
				| {
						dialog: true;
				  }
			) &
			(
				| {
						fullbleed: false;
						width: number;
				  }
				| {
						fullbleed: true;
				  }
			);

		export type Mount = Manifest.UnparsedMount;

		export type SoundCue = Manifest.UnparsedSoundCue;

		export type AssetCategory = Manifest.UnparsedAssetCategory;

		export type BundleDependencies = Manifest.UnparsedBundleDependencies;

		export type UnknownConfig = Record<string, unknown>;
	}

	export interface MooncgBundleConfig {
		databaseAdapter?: DatabaseAdapter;
	}

	/**
	 * The actual type of a `bundle`.
	 */
	export interface Bundle {
		name: string;
		version: string;
		license?: string;
		description?: string;
		homepage?: string;
		author?: Person;
		contributors?: Person[];
		dir: string;
		git: Bundle.GitData;
		transformBareModuleSpecifiers: boolean;
		hasAssignableSoundCues: boolean;
		hasExtension: boolean;
		config: Bundle.UnknownConfig;
		dashboard: {
			dir: string;
			panels: Bundle.Panel[];
		};
		graphics: Bundle.Graphic[];
		assetCategories: Bundle.AssetCategory[];
		mount: Bundle.Mount[];
		soundCues: Bundle.SoundCue[];
		compatibleRange?: string;
		bundleDependencies?: Bundle.BundleDependencies;
		mooncgBundleConfig: MooncgBundleConfig;
	}
// ...
}